export the BOLT_PROJECT env variable
Since bolt 2.18 the job fails with this error
```
Project directory '/builds/gfa/puppet-bolt' is world-writable which poses a security risk.
Set BOLT_PROJECT='/builds/gfa/puppet-bolt' to force the use of this project directory.
``

The permissions on the project directory don't matter for this case (a
CI job) I set the variable bolt wants